requirements.txt
boto3
Pillow
SQLAlchemy
psycopg2-binary
python-dotenv
PyYAML